refactor: clarify candidate module boundaries
