+++ src/evolve/candidate_smoke.py
