+++ tests/test_coherence.py
@@ -19,8 +19,7 @@ APPROVED_MODULES = {
     "__main__.py",
     "agent.py",
     "archive.py",
-    "asset_discovery.py",
-    "candidate_runtime.py",
+    "candidate_smoke.py",
     "candidate_snapshot.py",
     "cli.py",
     "config.py",
